Move utils folder outside of components folder
